+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Changed reaction emoji sorting [PR #391](https://github.com/marmot-protocol/whitenoise/pull/391)
 
 ### Deprecated
 
 ### Removed
+
 
 ### Fixed